implement inode read/write logic

+++ src/bitmap.c
@@ -1,14 +1,14 @@
 #include <sigmafs-tools/fs.h>
 
 
-int bitarr_write_bit(struct dev *dev, uint64_t block, uint32_t offset, uint8_t res)
+int bitarr_write_bit(struct dev *dev, uint32_t block, uint32_t offset, uint8_t res)
 {
 	if (!dev)
 		return 1;
-	uint64_t byte_offset = offset / 8;
-	uint64_t bit = offset % 8;
-	uint64_t block_num = block + (byte_offset / dev->block_size);
-	uint64_t block_offset = byte_offset % dev->block_size;
+	uint32_t byte_offset = offset / 8;
+	uint32_t bit = offset % 8;
+	uint32_t block_num = block + (byte_offset / dev->block_size);
+	uint32_t block_offset = byte_offset % dev->block_size;
 	uint8_t *buf = malloc(dev->block_size);
 	if (!buf)
 		return 1;
@@ -28,14 +28,14 @@ int bitarr_write_bit(struct dev *dev, uint64_t block, uint32_t offset, uint8_t r
 	return 0;
 }
 
-int bitarr_read_bit(struct dev *dev, uint64_t block, uint32_t offset, uint8_t *res)
+int bitarr_read_bit(struct dev *dev, uint32_t block, uint32_t offset, uint8_t *res)
 {
 	if (!dev || !res) 
 		return 1; 
-	uint64_t byte_offset = offset / 8; 
-	uint64_t bit = offset % 8; 
-	uint64_t block_num = block + (byte_offset / dev->block_size); 
-	uint64_t block_offset = byte_offset % dev->block_size; 
+	uint32_t byte_offset = offset / 8; 
+	uint32_t bit = offset % 8; 
+	uint32_t block_num = block + (byte_offset / dev->block_size); 
+	uint32_t block_offset = byte_offset % dev->block_size; 
 	uint8_t *buf = malloc(dev->block_size); 
 	if (!buf) 
 		return 1;
